Move utility functions to separate directory
